+++ GHCollection/Networking/Search/SearchSortBy.swift
@@ -1,0 +1,16 @@
+//
+//  SearchSortBy.swift
+//  GHCollection
+//
+//  Created by Piotr Bogdan on 29/05/2019.
+//  Copyright © 2019 Piotr Bogdan. All rights reserved.
+//
+
+import UIKit
+
+enum SearchSortBy: String {
+    case stars
+    case forks
+    case updated
+    case helpWantedIssues = "help-wanted-issues"
+}